Shifting to Front-End for a Bit
After several months of focusing on pure JS, I was getting very frustrated because I was making little to no progress.

I decided to switch things up a bit, and go back to front-end concepts, and I am glad I did. I am now in love with VSCode, and a month long break from HTML and CSS has really given me a fresh perspective on their concepts. 

This has a long way to go, but I want to code Nerdy Pug's entire website myself, this is the first of many commits on that project.